Key eligible-shipping-methods query on order.updatedAt
Every draft-order mutation that can affect shipping eligibility already
calls refreshEntity(), which bumps order.updatedAt. Keying the
eligible-shipping-methods query on [orderId, updatedAt] makes it
refetch automatically instead of maintaining a manual
invalidateQueries() call per mutation site -- which had missed the
coupon-code mutations entirely.

Replaces the regression test with one in sales/orders.spec.ts that
reproduces the actual stale-cache path (set customer first, enabling
the query on a zero-line order, then add a product line) against the
real Card-grid UI, per review on #5253.

Relates to #5253
