+++ packages/dashboard/src/app/routes/_authenticated/_orders/orders_.draft.$id.tsx
@@ -17,7 +17,7 @@ import { ActionBarItem } from '@/vdb/framework/layout-engine/action-bar-item-wra
 import { useDetailPage } from '@/vdb/framework/page/use-detail-page.js';
 import { api } from '@/vdb/graphql/api.js';
 import { Trans, useLingui } from '@lingui/react/macro';
-import { useMutation, useQuery, useQueryClient } from '@tanstack/react-query';
+import { useMutation, useQuery } from '@tanstack/react-query';
 import { createFileRoute, Link, useNavigate } from '@tanstack/react-router';
 import { ResultOf } from 'gql.tada';
 import { User } from 'lucide-react';
@@ -98,26 +98,21 @@ function DraftOrderPage() {
         },
     });
 
-    const queryClient = useQueryClient();
-    const eligibleShippingMethodsQueryKey = ['eligibleShippingMethods', entity?.id];
-
+    // Eligible shipping methods depend on the order's lines, customer, coupons,
+    // and shipping address -- any of which can change on any draft-order
+    // mutation, all of which already call refreshEntity(). Keying the query on
+    // entity.updatedAt (bumped by every such mutation) refetches automatically
+    // instead of maintaining a separate invalidation call per mutation, which
+    // previously missed coupon changes and left the list stale until reload (#5253).
     const { data: eligibleShippingMethods } = useQuery({
-        queryKey: eligibleShippingMethodsQueryKey,
+        queryKey: ['eligibleShippingMethods', entity?.id, entity?.updatedAt],
         queryFn: () => api.query(draftOrderEligibleShippingMethodsDocument, { orderId: entity?.id ?? '' }),
         enabled: !!entity?.shippingAddress?.streetLine1,
-        // The queryKey carries the order's identity, so keepPreviousData
-        // (the app-wide default) would show a previous draft order's
-        // shipping methods while switching between orders. Opt out.
+        // The queryKey carries the order's identity and version, so
+        // keepPreviousData (the app-wide default) would show a stale or
+        // previous order's shipping methods while the new list loads. Opt out.
         placeholderData: undefined,
     });
-
-    // Eligible shipping methods depend on the order's lines, customer, and
-    // shipping address -- invalidate after any mutation that can change one
-    // of those, alongside the existing refreshEntity() calls, otherwise the
-    // dashboard keeps showing a stale list until the page is reloaded (#5253).
-    const invalidateEligibleShippingMethods = () => {
-        queryClient.invalidateQueries({ queryKey: eligibleShippingMethodsQueryKey });
-    };
 
     const { mutate: addItemToDraftOrder } = useMutation({
         mutationFn: api.mutate(addItemToDraftOrderDocument),
@@ -127,7 +122,6 @@ function DraftOrderPage() {
                 case 'Order':
                     toast.success(t`Item added to order`);
                     refreshEntity();
-                    invalidateEligibleShippingMethods();
                     break;
                 default:
                     toast.error(order.message);
@@ -149,7 +143,6 @@ function DraftOrderPage() {
                 case 'Order':
                     toast.success(t`Order line updated`);
                     refreshEntity();
-                    invalidateEligibleShippingMethods();
                     break;
                 default:
                     toast.error(order.message);
@@ -169,7 +162,6 @@ function DraftOrderPage() {
                 case 'Order':
                     toast.success(t`Order line removed`);
                     refreshEntity();
-                    invalidateEligibleShippingMethods();
                     break;
                 default:
                     toast.error(order.message);
@@ -235,7 +227,6 @@ function DraftOrderPage() {
                         toast.error(t`Failed to set address for order: ${e instanceof Error ? e.message : String(e)}`);
                     }
                     refreshEntity();
-                    invalidateEligibleShippingMethods();
                     break;
                 }
                 default:
@@ -253,7 +244,6 @@ function DraftOrderPage() {
         onSuccess: (result: ResultOf<typeof setShippingAddressForDraftOrderDocument>) => {
             toast.success(t`Shipping address set for order`);
             refreshEntity();
-            invalidateEligibleShippingMethods();
         },
         onError: error => {
             toast.error(t`Failed to set shipping address for order: ${error.message}`);
@@ -276,7 +266,6 @@ function DraftOrderPage() {
         onSuccess: (result: ResultOf<typeof unsetShippingAddressForDraftOrderDocument>) => {
             toast.success(t`Shipping address unset for order`);
             refreshEntity();
-            invalidateEligibleShippingMethods();
         },
         onError: error => {
             toast.error(t`Failed to unset shipping address for order: ${error.message}`);
